code/localOcrPaddle/requirements.txt
flask==3.0.3
flask-cors==4.0.1
numpy==1.26.4
pillow==10.4.0
opencv-python-headless==4.10.0.84
paddleocr==2.7.3
# CPU PaddlePaddle; pinned to a commonly available build.
# If this fails to install for your platform, adjust version in this file.
paddlepaddle==2.6.2
